Added many files (n deleted 5) deemed almost complete to be committed!
Note: The files which I deleted were done so for a reason I do not really remember anymore entirely however I believe it was because I had to delete them in order to fix some bug I was having when just even trying to import Keras (raised various weird errors sometimes to do with NumPy). However what had actually fixed this issue mentioned  was the use of another environment I had created called AlgoTradingEnv2! The scripts called dojo and testing1 however were deleted for " deleted for "tidying" purposes!
In addition to this I also added scripts which I consider to be completed enough or necessary enough in their current states (but which are works in progress) for committing!